+++ test/endpoint.js
@@ -43,6 +43,7 @@ describe('clustering', function() {
     require('./clustering/record-linkage/helpers.js');
 
     require('./clustering/record-linkage/blocking.js');
+    require('./clustering/record-linkage/canopy.js');
     require('./clustering/record-linkage/key-collision.js');
     require('./clustering/record-linkage/naive.js');
     require('./clustering/record-linkage/sorted-neighborhood.js');